Implement search input functionality
Search input now works correctly and displays a message if the searched product does not exist.
When a user searches while in the cart, they are redirected to the products page to view results.
The search input is also cleared automatically when a category button is clicked while the input has content.